+++ src/persistencia/dao/interfaz/DAO.java
@@ -7,7 +7,7 @@ public interface DAO<T>
 	public T create(T t);
 	public void delete(T t);
 	public T update(T t);
-	public T byId(long id);	
+	public T byId(int id);	
 	public List<T> all();
 	public void close();
 }
